Linux/GCC 64bit build fixes + PNG callback type fix
git-svn-id: svn://libnui.net/nui/trunk/nui3@4685 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ src/Dialogs/nuiMessageBox.cpp
@@ -252,7 +252,7 @@ bool nuiMessageBox::OnRetry(const nuiEvent& rEvent)
 
 bool nuiMessageBox::OnCustom(const nuiEvent& rEvent)
 {
-  Button btn = (Button)rEvent.mpUser; 
+  Button btn = (Button)(unsigned long int)rEvent.mpUser; 
   mClickedButton = btn;
   ExitModal();
   return false;
